Add WALKTHROUGH_LOCATIONS param and env to the web app template
There has been a change to the web app to use an environment var
for the locations of walkthroughs, instead of bundling them in with
the web app. This env var is called WALKTHROUGH_LOCATIONS.

If the env var is not set then there is a default value. However, we
should allow this value to be overridden in the operator.

This change allows the env var to be set by setting the
WALKTHROUGH_LOCATIONS param in the template.