realign template and internal var variables
realign template and internal var variables and ensure that the
internal variable is appended when processing the template in the
future.

this will ensure that there is one source of truth for the template
version.

the default version in the template should still be updated, however
it is no longer necessary to avoid an installation having to perform
a reconcile of the image if the two image versions are not aligned.